revert restructure api
