+++ app/__init__.py
@@ -2,9 +2,8 @@ from flask import Flask
 
 # routes
 from app.routes.main import main
+from app.routes.api import api
 from app.routes.auth import auth
-from app.routes.api.v1 import events
-
 
 # db instance
 from app.models import db
@@ -21,7 +20,7 @@ def create_app(configuration):
 
 	# register blueprints
 	app.register_blueprint(main)
+	app.register_blueprint(api, url_prefix=app.config['API_BASE_URL'])
 	app.register_blueprint(auth, url_prefix='/auth')
-	app.register_blueprint(events, url_prefix=app.config['API_BASE_URL']+'/events')
-	
+
 	return app
