--- app/routes/api/v1/__init__.py
@@ -1,7 +1,0 @@
-'''
-put api route in here
-'''
-
-""" import all api related routes here"""
-from app.routes.api.v1.events import events
-
