+++ seeders/events_seeder.py
@@ -7,6 +7,7 @@ from app.models import db
 Seeder class for
 '''
 
+
 class EventsSeeder():
 
     @staticmethod
@@ -18,13 +19,14 @@ class EventsSeeder():
         for i in range(0, 10):
             time_start = fake.future_datetime(end_date="+30d", tzinfo=None)
             time_end = time_start + timedelta(hours=3)
-            information = fake.sentence(nb_words=6, variable_nb_words=True, ext_word_list=None)
+            information = fake.sentence(
+                nb_words=6, variable_nb_words=True, ext_word_list=None)
             event = {
                 'information': information,
                 'time_start': time_start,
                 'time_end': time_end
             }
-            
+
             new_event = Events()
             new_event.information = event['information']
             new_event.time_start = event['time_start']
